+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- chore: add SwiftUI View extensions to capture screen view and views in general (postHogViewEvent, postHogScreenView) ([#180](https://github.com/PostHog/posthog-ios/pull/180))
+
 ## 3.8.3 - 2024-09-03
 
 - recording: use non deprecated methods for getCurrentWindow if available ([#178](https://github.com/PostHog/posthog-ios/pull/178))